requirements.txt
pandas==1.1.3
folium==0.11.0
geopandas==0.9.0
streamlit==0.83.0
numpy==1.20.1
plotly==5.0.0
pyodbc==4.0.30
tqdm==4.61.1
streamlit_folium==0.4.0
geopy==2.1.0
matplotlib==3.4.1
Pillow==8.2.0
